Updated gitignore. Removed old matrix class. Add WoR-World and WoR-robot documentation. Added some extra compile flags for robotapplication. Fixed some compile warnings. Fixed some cppcheck errors. Updated robotapplication README.

+++ robotvision/src/ApplicationManager.cpp
@@ -84,7 +84,7 @@ void ApplicationManager::specialCommandParser(std::string& command)
     {
         shapeFinder.setRealLifeConversionRate(
             colorFinder.applyFilter(inputHandler.getVideoCaptureFrame(), ColorFilter::Color::WHITE, showFilterProgress),
-            std::stoi(match[1]), showFilterProgress);
+            std::stof(match[1]), showFilterProgress);
         return;
     }
     if (command.find("applyfilter") != std::string::npos)
